First Game build
here is the current state of the game right now in a playable format. To close the game you still need to force close it with either task manager or using Alt + F4

there is a start menu and a pause menu though, so if needed it can be paused, i do think i need to increase the speed of the characters, cause in the build they seem kinda slow. there will eventually be a instructions page before you move on and some background music